+++ xdef/src/main/java/org/xdef/XDConstants.java
@@ -106,8 +106,7 @@ public interface XDConstants {
 	/** URI of w3c XLink */
 	public static final String XLINK_NS_URI = "http://www.w3.org/1999/xlink";
 	/** URI of w3c XInclude */
-	public static final String XINCLUDE_NS_URI =
-		"http://www.w3.org/2001/XInclude";
+	public static final String XINCLUDE_NS_URI = "http://www.w3.org/2001/XInclude";
 
 ////////////////////////////////////////////////////////////////////////////////
 // X-definition
@@ -127,8 +126,7 @@ public interface XDConstants {
 	public static final String XDEF42_NS_URI = "http://www.xdef.org/xdef/4.2";
 
 	/** The namespace URI for X-definition instance. */
-	public static final String XDEF_INSTANCE_NS_URI =
-		"http://www.xdef.org/xdef/instance";
+	public static final String XDEF_INSTANCE_NS_URI = "http://www.xdef.org/xdef/instance";
 
 ////////////////////////////////////////////////////////////////////////////////
 // XON/JSON/INI
@@ -155,6 +153,5 @@ public interface XDConstants {
 	/** Build version of software build from pom.xml */
 	public static final String BUILD_VERSION = PomInfo.POMINFO.getVersion();
 	/** Datetime of build version from pom.xml */
-	public static final String BUILD_DATETIME =
-		PomInfo.POMINFO.getBuildTimestamp();
+	public static final String BUILD_DATETIME = PomInfo.POMINFO.getBuildTimestamp();
 }