+++ xdef/src/main/java/org/xdef/XDDocument.java
@@ -103,17 +103,17 @@ public interface XDDocument extends XXNode {
 	 */
 	public void setRootModel(XMElement xmodel);
 
-	/** get StdOut.
+	/** get StdOut output.
 	 * @return std out XDOutput.
 	 */
 	public XDOutput getStdOut();
 
-	/** get StdErr.
+	/** get StdErr output.
 	 * @return std err XDOutput.
 	 */
 	public XDOutput getStdErr();
 
-	/** get StdIn.
+	/** get StdIn intput.
 	 * @return std in XDInput.
 	 */
 	public XDInput getStdIn();
@@ -121,8 +121,7 @@ public interface XDDocument extends XXNode {
 	/** Set XML writer.
 	 * @param out output stream.
 	 * @param encoding encoding of output.
-	 * @param writeDocumentHeader if true full document is written, otherwise
-	 * only root element.
+	 * @param writeDocumentHeader if true full document is written, otherwise only root element.
 	 * @throws IOException if an error occurs.
 	 */
 	public void setStreamWriter(OutputStream out,
@@ -132,8 +131,7 @@ public interface XDDocument extends XXNode {
 	/** Set XML writer.
 	 * @param out stream writer.
 	 * @param encoding encoding of output.
-	 * @param writeDocumentHeader if true full document is written, otherwise
-	 * only root element.
+	 * @param writeDocumentHeader if true full document is written, otherwise only root element.
 	 */
 	public void setStreamWriter(Writer out,
 		String encoding,
@@ -196,8 +194,8 @@ public interface XDDocument extends XXNode {
 	 * @param nsURI Namespace URI of the element.
 	 * @param qname Qualified name of the element.
 	 * @param checkRoot If value of this argument is true then the root check
-	 * element is checked against the root list, otherwise it is found as
-	 * XElement on the base level.
+	 * element is checked against the root list, otherwise it is found
+	 * as XElement on the base level.
 	 * @return ChkElement object.
 	 */
 	public XXElement prepareRootXXElementNS(final String nsURI,
@@ -207,8 +205,8 @@ public interface XDDocument extends XXNode {
 	/** Create root check element with given name.
 	 * @param name Tag name of the root element.
 	 * @param checkRoot If value of this argument is true then the root check
-	 * element is checked against the root list, otherwise it is found as
-	 * XElement on the base level.
+	 * element is checked against the root list, otherwise it is found
+	 * as XElement on the base level.
 	 * @return ChkElement object.
 	 */
 	public XXElement prepareRootXXElement(final String name, boolean checkRoot);
@@ -281,15 +279,13 @@ public interface XDDocument extends XXNode {
 
 	/** Set actual source language used for lexicon.
 	 * @param language string with language or null.
-	 * @throws SRuntimeException if lexicon not specified
-	 * or if language is not specified.
+	 * @throws SRuntimeException if lexicon not specified or if language is not specified.
 	 */
 	public void setLexiconLanguage(String language) throws SRuntimeException;
 
 	/** Translate the input element from the source language to the destination
 	 * language according to lexicon.
-	 * @param elem path to the source element or the string
-	 * with element.
+	 * @param elem path to the source element or the string with element.
 	 * @param sourceLanguage name of source language.
 	 * @param destLanguage name of destination language.
 	 * @param reporter the reporter where to write errors or null.
@@ -399,10 +395,8 @@ public interface XDDocument extends XXNode {
 		throws SRuntimeException;
 
 	/** Parse source INI/Properties and return XComponent as result.
-	 * @param ini string with pathname of INI/Properties file
-	 * or INI/Properties source data.
-	 * @param xClass XComponent class (if null, then XComponent class
-	 * is searched in XDPool).
+	 * @param ini string with pathname of INI/Properties file or INI/Properties source data.
+	 * @param xClass XComponent class (if null, then XComponent class is searched in XDPool).
 	 * @param reporter report writer or null. If this argument is
 	 * null and error reports occurs then SRuntimeException is thrown.
 	 * @return XComponent with parsed data.
@@ -413,10 +407,8 @@ public interface XDDocument extends XXNode {
 		ReportWriter reporter) throws SRuntimeException;
 
 	/** Parse source INI/Properties and return XComponent as result.
-	 * @param ini string with pathname of INI/Properties file
-	 * or INI/Properties source data.
-	 * @param xClass XComponent class (if null, then XComponent class
-	 * is searched in XDPool).
+	 * @param ini string with pathname of INI/Properties file or INI/Properties source data.
+	 * @param xClass XComponent class (if null, then XComponent class is searched in XDPool).
 	 * @param sourceId name of source or null.
 	 * @param reporter report writer or null. If this argument is
 	 * null and error reports occurs then SRuntimeException is thrown.
@@ -429,8 +421,7 @@ public interface XDDocument extends XXNode {
 		ReportWriter reporter) throws SRuntimeException;
 
 	/** Validate and process INI/Properties data and return processed object.
-	 * @param data INI/Properties object or XML representation of object
-	 * to validate.
+	 * @param data INI/Properties object or XML representation of object to validate.
 	 * @param reporter report writer or null. If this argument is
 	 * null and error reports occurs then SRuntimeException is thrown.
 	 * @return Map with processed data.
@@ -452,7 +443,7 @@ public interface XDDocument extends XXNode {
 	/** Create XComponent from XON/JSON according to the X-definition model.
 	 * NOTE this method is experimental.
 	 * @param name the name of required model.
-	 * @param xClass XComponent class (if <i>null</i>, then XComponent class
+	 * @param xClass XComponent class (if <i>null</i>, then XComponent class.
 	 * @param reporter report writer or <i>null</i>. If this argument is
 	 * <i>null</i> and error reports occurs then SRuntimeException is thrown.
 	 * @return XComponent with created data.
@@ -504,8 +495,7 @@ public interface XDDocument extends XXNode {
 
 	/** Parse source XON/JSON and return XComponent as result.
 	 * @param data string with pathname of XON/JSON source data.
-	 * @param xClass XComponent class (if null, then XComponent class
-	 * is searched in XDPool).
+	 * @param xClass XComponent class (if null, then XComponent class is searched in XDPool).
 	 * @param reporter report writer or null. If this argument is
 	 * null and error reports occurs then SRuntimeException is thrown.
 	 * @return XComponent with parsed data.
@@ -517,8 +507,7 @@ public interface XDDocument extends XXNode {
 
 	/** Parse source JSON/XON and return XComponent as result.
 	 * @param data string with pathname of XON/JSON source data.
-	 * @param xClass XComponent class (if null, then XComponent class
-	 * is searched in XDPool).
+	 * @param xClass XComponent class (if null, then XComponent class is searched in XDPool).
 	 * @param sourceId name of source or null.
 	 * @param reporter report writer or null. If this argument is
 	 * null and error reports occurs then SRuntimeException is thrown.
@@ -551,9 +540,8 @@ public interface XDDocument extends XXNode {
 		throws SRuntimeException;
 
 	/** Run create mode - create element according to the X-definition model.
-	 * If the parameter nsUri is not null then its assigned the model
-	 * with given namespaceURI; in this case the parameter name may be
-	 * qualified with a prefix.
+	 * If the parameter nsUri is not null then its assigned the model with given namespaceURI;
+	 * in this case the parameter name may be qualified with a prefix.
 	 * @param nsUri the namespace URI of result element (may be null).
 	 * @param name the name of model of required element (may contain prefix).
 	 * @param reporter report writer or null. If this argument is
@@ -599,8 +587,7 @@ public interface XDDocument extends XXNode {
 
 	/** Parse source XML and return XComponent as result.
 	 * @param data string with pathname of XML file or XML source data.
-	 * @param xClass XComponent class (if null, then XComponent class
-	 * is searched in XDPool).
+	 * @param xClass XComponent class (if null, then XComponent class is searched in XDPool).
 	 * @param reporter report writer or null. If this argument is
 	 * null and error reports occurs then SRuntimeException is thrown.
 	 * @return XComponent with parsed data.
@@ -612,8 +599,7 @@ public interface XDDocument extends XXNode {
 
 	/** Parse source XML and return XComponent as result.
 	 * @param data string with pathname of XML file or XML source data.
-	 * @param xClass XComponent class (if null, then XComponent class
-	 * is searched in XDPool).
+	 * @param xClass XComponent class (if null, then XComponent class is searched in XDPool).
 	 * @param sourceId name of source or null.
 	 * @param reporter report writer or null. If this argument is
 	 * null and error reports occurs then SRuntimeException is thrown.
@@ -627,8 +613,7 @@ public interface XDDocument extends XXNode {
 
 	/** Run create mode - create element according to the X-definition model.
 	 * If the parameter nsUri is not <i>null</i> then its assigned the model
-	 * with given namespaceURI; in this case the parameter qname may be
-	 * qualified with a prefix.
+	 * with given namespaceURI (in this case the parameter qname may be qualified with a prefix).
 	 * @param nsUri the namespace URI of result element (may be <i>null</i>).
 	 * @param name the name of model of required element (may contain prefix).
 	 * @param xClass XComponent class (if <i>null</i>, then XComponent class
@@ -644,7 +629,7 @@ public interface XDDocument extends XXNode {
 
 	/** Run create mode - create XComponent according to the X-definition model.
 	 * @param name the name of model of required element.
-	 * @param xClass XComponent class (if <i>null</i>, then XComponent class
+	 * @param xClass XComponent class (if <i>null</i>, then XComponent class.
 	 * @param reporter report writer or <i>null</i>. If this argument is
 	 * <i>null</i> and error reports occurs then SRuntimeException is thrown.
 	 * @return XComponent with created data.
@@ -656,12 +641,11 @@ public interface XDDocument extends XXNode {
 
 	/** Run create mode - create XComponent according to the X-definition model.
 	 * @param qname the QName of model of required element.
-	 * @param xClass XComponent class (if <i>null</i>, then XComponent class
+	 * @param xClass XComponent class (if <i>null</i>, then XComponent class.
 	 * @param reporter report writer or <i>null</i>. If this argument is
 	 * <i>null</i> and error reports occurs then SRuntimeException is thrown.
 	 * @return XComponent with created data.
-	 * @throws SRuntimeException if reporter is <i>null</i> and an error
-	 * was reported.
+	 * @throws SRuntimeException if reporter is <i>null</i> and an error was reported.
 	 */
 	public XComponent xcreateXComponent(final QName qname,
 		final Class xClass,
@@ -732,8 +716,7 @@ public interface XDDocument extends XXNode {
 
 	/** Parse source YAML and return XComponent as result.
 	 * @param data string with pathname of YAML file or YAML source data.
-	 * @param xClass XComponent class (if null, then XComponent class
-	 * is searched in XDPool).
+	 * @param xClass XComponent class (if null, then XComponent class is searched in XDPool).
 	 * @param sourceId name of source or null.
 	 * @param reporter report writer or null. If this argument is
 	 * null and error reports occurs then SRuntimeException is thrown.
@@ -750,8 +733,7 @@ public interface XDDocument extends XXNode {
 	/** Parse source XML and return XComponent as result.
 	 * @deprecated please use xparseXComponent instead
 	 * @param data string with pathname of XML file or XML source data.
-	 * @param xClass XComponent class (if null, then XComponent class
-	 * is searched in XDPool).
+	 * @param xClass XComponent class (if null, then XComponent class is searched in XDPool).
 	 * @param reporter report writer or null. If this argument is
 	 * null and error reports occurs then SRuntimeException is thrown.
 	 * @return XComponent with parsed data.
@@ -764,8 +746,7 @@ public interface XDDocument extends XXNode {
 	/** Parse source XML and return XComponent as result.
 	 * @deprecated please use xparseXComponent instead
 	 * @param data string with pathname of XML file or XML source data.
-	 * @param xClass XComponent class (if null, then XComponent class
-	 * is searched in XDPool).
+	 * @param xClass XComponent class (if null, then XComponent class is searched in XDPool).
 	 * @param sourceId name of source or null.
 	 * @param reporter report writer or null. If this argument is
 	 * null and error reports occurs then SRuntimeException is thrown.