+++ xdef/src/main/java/org/xdef/XDElementAbstract.java
@@ -24,7 +24,7 @@ import static org.xdef.XDValueType.ELEMENT;
 import org.xdef.sys.SUnsupportedOperationException;
 import org.xdef.xml.KXmlUtils;
 
-/** Abstract class for implementation of "pseudo" elements.
+/** Abstract class for implementation of "pseudo" elements  in X-script.
  * @author Vaclav Trojan
  */
 public abstract class XDElementAbstract extends XDValueAbstract